refactor(paste): split menu fallback executor into focused components
- extract menu traversal/discovery into pastemenuscanner and move live ax observer logic into pasteaxlivesession, keeping pastemenufallbackexecutor orchestration-only.

- wire new files into xcode project and update codemap/engineering docs. no policy or behavior changes intended (structural refactor only).
